Revert "Moved windows test into IntegrationTests"
This reverts commit 5f81cbf9aa30b750ccfff848bed1cb59cf786232.

# Conflicts:
#	Packages/com.unity.inputsystem/Tests/IntegrationTests/IntegrationTests.cs
