--- Packages/com.unity.inputsystem/TestRunnerOptions.json
@@ -1,3 +1,0 @@
-{
-    "disableBatchMode": true
-}
