+++ Packages/com.unity.inputsystem/Tests/IntegrationTests/IntegrationTests.cs
@@ -1,9 +1,6 @@
-using System.Collections;
 using NUnit.Framework;
-using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.InputSystem.LowLevel;
-using UnityEngine.TestTools;
 
 // Disable irrelevant warning about there not being underscores in method names.
 #pragma warning disable CA1707
@@ -43,50 +40,4 @@ public class IntegrationTests
             InputSystem.RemoveDevice(keyboard);
         }
     }
-
-    #if UNITY_2019_3_OR_NEWER && (UNITY_STANDALONE_WIN || UNITY_EDITOR_WIN)
-
-    [UnityTest]
-    [Category("Integration")]
-    public IEnumerator WindowsInput_RemoteDesktopMouseMovements_AreDetected()
-    {
-        var mouse = InputSystem.GetDevice<Mouse>();
-        var currentPosition = mouse.position.ReadValue();
-
-        yield return new WaitForSeconds(0.1f);
-        Assert.AreEqual(currentPosition, mouse.position.ReadValue(), "Expected mouse position to not change when no input was sent. Please do not move the mouse during this test.");
-
-        WinUserInput.SendRDPMouseMoveEvent(10, 10);
-        yield return new WaitForSeconds(0.1f);
-
-        Assert.AreNotEqual(currentPosition, mouse.position.ReadValue(), "Expected mouse position to have moved when sending RDP/absolute values.");
-        currentPosition = mouse.position.ReadValue();
-
-        WinUserInput.SendRDPMouseMoveEvent(100, 100);
-        yield return new WaitForSeconds(0.1f);
-        Assert.AreNotEqual(currentPosition, mouse.position.ReadValue(), "Expected mouse position to have moved when sending RDP/absolute values.");
-    }
-
-    [UnityTest]
-    [Category("Integration")]
-    public IEnumerator WindowsInput_MouseMovements_AreDetected()
-    {
-        var mouse = InputSystem.GetDevice<Mouse>();
-        var currentPosition = mouse.position.ReadValue();
-
-        yield return new WaitForSeconds(0.1f);
-        Assert.AreEqual(currentPosition, mouse.position.ReadValue(), "Expected mouse position to not change when no input was sent. Please do not move the mouse during this test.");
-
-        WinUserInput.SendMouseMoveEvent(10, 10);
-        yield return new WaitForSeconds(0.1f);
-
-        Assert.AreNotEqual(currentPosition, mouse.position.ReadValue(), "Expected mouse position to have moved when sending relative values.");
-        currentPosition = mouse.position.ReadValue();
-
-        WinUserInput.SendMouseMoveEvent(100, 100);
-        yield return new WaitForSeconds(0.1f);
-        Assert.AreNotEqual(currentPosition, mouse.position.ReadValue(), "Expected mouse position to have moved when sending relative values.");
-    }
-
-    #endif // UNITY_2019_3_OR_NEWER && (UNITY_STANDALONE_WIN || UNITY_EDITOR_WIN)
 }
